Update index file for rootReducer to tsx with appropriate TS
